Rename UdongCheckbox filename
